cleanReleases added
