+++ apg-patch-service-cmdclient/src/main/resources/apscli.properties
@@ -1,6 +1,7 @@
 environments {
     test {
         revision.file.path = 'src/test/resources/Revisions.json'
+        revision.range.step = 10000
         host.default = 'localhost:9020'
         config.dir = 'src/test/resources/config'
         onclone.delete.artifact.dryrun = true
@@ -11,6 +12,7 @@ environments {
     }
     production {
         revision.file.path = '/var/opt/apg-patch-cli/Revisions.json'
+        revision.range.step = 10000
         host.default = 'localhost:9010'
         config.dir = '/var/opt/apg-patch-common'
         onclone.delete.artifact.dryrun = false
